+++ src/common/extensions/security.ts
