+++ src/common/extensions/string.ts
